refactor(web): rename `keyboard.ts` to `jsKeyboard.ts`
