+++ web/src/engine/keyboard/src/keyboards/activeLayout.ts
@@ -3,7 +3,7 @@ import KeyEvent, { KeyEventSpec } from "../keyEvent.js";
 import KeyMapping from "../keyMapping.js";
 import { ButtonClasses, Layouts } from "./defaultLayouts.js";
 import type { LayoutKey, LayoutSubKey, LayoutRow, LayoutLayer, LayoutFormFactor, ButtonClass } from "./defaultLayouts.js";
-import { type JSKeyboard } from "./keyboard.js";
+import { type JSKeyboard } from "./jsKeyboard.js";
 
 import { TouchLayout } from "@keymanapp/common-types";
 import TouchLayoutDefaultHint = TouchLayout.TouchLayoutDefaultHint;
